Collision detection & event broadcasting
This commit connects the different pieces of the game together through
events/delegates. PlayerCollisionDetector implements the
OnTriggerEnter2D method such that a `PlayerDeathEvent` is broadcasted to
the GameController, the player controls, and the moving
tiles/obstacles. This is a cleaner way of maintaining the state of the
game than we used in Rocky River.

This commit also contains some code cleanup/ reformatting.